Game Rule Architecture v0 modify 2 & demo
1. Modify RuleModel
- modify interface and class that it can design for space pattern.
2. DEMO_RM_SpacePattern_MatrixSpace
a. space is created by matrix, the matrix-space have the sub-space.
b. mouse address is globe address, this address need transform to space
address, and focus on sub-space in matrix-space.

3. DEMO_RM_SpacePattern_GeometrySpace
a. space is normal space, geometry-space don't have sub-space.
b. mouse address is globe address, detected need address and range, and
use Euclidean distance to check object is in range.