Game Rule Architecture v0 demo 2
1. DEMO_RM_SpacePattern_MatrixSpace_Keyboard_Control
a. The application of MatrixSpace, use keyboard( W, A, S, D ) to control
an AvatorActor.
b. The application is focus on topic "how to use actor control  or use
space information."

2. DEMO_RM_SpacePattern_MatrixSpace_Drag_Control
a. The application of MatrixSpace, use mouse to control an AvatorActor,
and AvatorActor will lock ChessActor and darg it.
b. ChessActor responsibility is control or use space information.
c. The application is focus on topic "actor is control by user or actor,
space is control by actor."